requirements.txt
Flask
mysql-connector-python
pytest
pytest-cov
